Bug fixes to enable more tests to run, remove a few junk files
components/file_read_msg.rcc/file_read_msg.xml: allow settable message size
components/sym_fir_complex.rcc/sym_fir_complex.c: deal with ZLM
components/sym_fir_real.rcc/sym_fir_real.c: ditto
env/start.sh: forget about the ocpi subdir of OCPI_BASE_DIR, use exports
examples/fsk_modem_xml/fsk_modem.cxx: use file_read/write, not file_read/write_msg
runtime/util/ezxml/src/OcpiUtilEzxml.cxx: backport 32 bit vs 64 bit bug fix from sdp for ddstest
components/bias_vhdl1.hdl/{ocpi-obj93.cf,try,work-obj93.cf,xx.prj}: removed junk
hdl/assemblies/README.txt: removed obsolete junk

+++ runtime/util/ezxml/src/OcpiUtilEzxml.cxx
@@ -519,7 +519,7 @@ namespace OCPI {
 	  break;
 	// Allow hex values without a minus to include the sign bit
 	if (bits && !minus && s[0] == '0' && s[1] == 'x' && // if hex and not minus
-	    (uval & (1 << (bits-1)))) {                     //  if sign bit set
+	    (uval & ((uint64_t)1 << (bits-1)))) {                     //  if sign bit set
 	  if (bits == 64)
 	    val = (int64_t)uval;
 	  else if (uval & (~(uint64_t)0 << bits)) // if it has invalid high order bits
